+++ FirebaseFunctions/Sources/Internal/FunctionsContext.swift
@@ -22,11 +22,14 @@ internal class FunctionsContext: NSObject {
   let authToken: String?
   let fcmToken: String?
   let appCheckToken: String?
+  let limitedUseAppCheckToken: String?
 
-  init(authToken: String?, fcmToken: String?, appCheckToken: String?) {
+  init(authToken: String?, fcmToken: String?, appCheckToken: String?,
+       limitedUseAppCheckToken: String?) {
     self.authToken = authToken
     self.fcmToken = fcmToken
     self.appCheckToken = appCheckToken
+    self.limitedUseAppCheckToken = limitedUseAppCheckToken
   }
 }
 
@@ -48,12 +51,14 @@ internal class FunctionsContextProvider: NSObject {
 //
 //  }
 
-  internal func getContext(_ completion: @escaping ((FunctionsContext, Error?) -> Void)) {
+  internal func getContext(options: HTTPSCallableOptions? = nil,
+                           _ completion: @escaping ((FunctionsContext, Error?) -> Void)) {
     let dispatchGroup = DispatchGroup()
 
     var authToken: String?
     var appCheckToken: String?
     var error: Error?
+    var limitedUseAppCheckToken: String?
 
     if let auth = auth {
       dispatchGroup.enter()
@@ -68,19 +73,30 @@ internal class FunctionsContextProvider: NSObject {
     if let appCheck = appCheck {
       dispatchGroup.enter()
 
-      appCheck.getToken(forcingRefresh: false) { tokenResult in
-        // Send only valid token to functions.
-        if tokenResult.error == nil {
-          appCheckToken = tokenResult.token
+      if options?.requireLimitedUseAppCheckTokens == true {
+        appCheck.getLimitedUseToken? { tokenResult in
+          // Send only valid token to functions.
+          if tokenResult.error == nil {
+            limitedUseAppCheckToken = tokenResult.token
+          }
+          dispatchGroup.leave()
         }
-        dispatchGroup.leave()
+      } else {
+        appCheck.getToken(forcingRefresh: false) { tokenResult in
+          // Send only valid token to functions.
+          if tokenResult.error == nil {
+            appCheckToken = tokenResult.token
+          }
+          dispatchGroup.leave()
+        }
       }
     }
 
     dispatchGroup.notify(queue: .main) {
       let context = FunctionsContext(authToken: authToken,
                                      fcmToken: self.messaging?.fcmToken,
-                                     appCheckToken: appCheckToken)
+                                     appCheckToken: appCheckToken,
+                                     limitedUseAppCheckToken: limitedUseAppCheckToken)
       completion(context, error)
     }
   }