+++ Firestore/core/src/remote/remote_store.h
@@ -42,6 +42,10 @@ namespace local {
 class LocalStore;
 }  // namespace local
 
+namespace model {
+class AggregateField;
+}  // namespace model
+
 namespace remote {
 
 class ConnectivityMonitor;
@@ -194,8 +198,9 @@ class RemoteStore : public TargetMetadataProvider,
   absl::optional<local::TargetData> GetTargetDataForTarget(
       model::TargetId target_id) const override;
 
-  void RunCountQuery(const core::Query& query,
-                     api::CountQueryCallback&& result_callback);
+  void RunAggregateQuery(const core::Query& query,
+                         const std::vector<model::AggregateField>& aggregates,
+                         api::AggregateQueryCallback&& result_callback);
 
   void OnWatchStreamOpen() override;
   void OnWatchStreamChange(