+++ FirebaseFunctions/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 10.10.0
+- [fixed] Fixed potential memory leak of Functions instances. (#11248)
+- [added] Callable functions can now opt in to using limited-use App Check
+  tokens. (#11270)
+
 # 10.0.0
 - [fixed] Remove unnecessary and unused `encoder` and `decoder` parameters from
   `public func call(_ data: Request) async throws -> Response` API. (#9891)